Cleaned up all files
