+++ src/App.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { Navbar } from './components/Navbar';
-import Home from './components/pages/Home';
-import About from './components/pages/About';
-import Login from './components/pages/Login';
+import Home from './pages/Home';
+import About from './pages/About';
+import Login from './pages/Login';
 import { GoogleOAuthProvider } from '@react-oauth/google';
 
 function App() {
