+++ src/components/Button.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Button.css'
+import './styles/Button.css'
 
 // Define the props type
 interface ButtonProps {
