+++ src/components/Footer.tsx
@@ -1,6 +1,6 @@
 // src/components/Footer.tsx
 import React from 'react';
-import "./Footer.css";
+import "./styles/Footer.css";
 import Logo from './images/logo_edited.png'
 
 export const Footer = () => {
