+++ src/components/homeSections/HomeBlock3.tsx
@@ -1,7 +1,7 @@
 // src/components/Home.tsx
 import React from 'react';
 import "./styles/HomeBlock3.css";
-import VerticalBlockItem from '../../VerticalBlockItem';
+import VerticalBlockItem from '../VerticalBlockItem';
 import { FaMagnifyingGlass } from "react-icons/fa6";
 import { FaPhoneAlt } from "react-icons/fa";
 import { GiHouseKeys } from "react-icons/gi";
