+++ src/components/homeSections/images/dormRoomSample.png
