+++ src/components/homeSections/styles/HomeBlock1.css
