+++ src/components/homeSections/styles/HomeBlock2.css
